+++ explorer/src/lib/components/__tests__/BlocksCard.spec.js
@@ -62,4 +62,15 @@ describe("Blocks Card", () => {
 
     expect(container.firstChild).toMatchSnapshot();
   });
+
+  it("should hide the `Show More` button if an error has occurred", async () => {
+    const props = { ...baseProps, error: new Error("error") };
+
+    const { container } = render(BlocksCard, {
+      ...baseOptions,
+      props: { ...props, error: new Error("error") },
+    });
+
+    expect(container.firstChild).toMatchSnapshot();
+  });
 });